requirements.txt
joblib==1.2.0
matplotlib==3.8.0
pandas==2.2.1
Pillow==10.2.0
plotly==5.9.0
Requests==2.31.0
seaborn==0.13.2
streamlit==1.32.2

